Fix: resolves issue #1296 (#1313)
The /logos page is not loading properly as /logos?logo_ids=4289009%2C5413042 returns 404 bad request. This happened as the backend expects it to be in the form /logos?logo_ids=4289009%2Clogo_ids=5413042. The getLogosImages() was returning request status of 400. Changing the api request params format from frontend solves the issue